+++ packages/data-schema/src/runtime/client/index.ts
@@ -48,7 +48,7 @@ type Model = Record<string, any>;
 /**
  * Currently this omits any object-type fields. Update this when we add custom types/enums.
  */
-type NonRelationalFields<M extends Model> = {
+type NonRelationshipFields<M extends Model> = {
   [Field in keyof M as UnwrapArray<M[Field]> extends Record<string, unknown>
     ? never
     : Field]: M[Field];
@@ -166,7 +166,11 @@ type DeepPickFromPath<
       : Path extends `${infer Head}.${infer Tail}`
         ? Head extends keyof FlatModel
           ? Tail extends '*'
-            ? { [k in Head]: NonRelationalFields<UnwrapArray<FlatModel[Head]>> }
+            ? {
+                [k in Head]: NonRelationshipFields<
+                  UnwrapArray<FlatModel[Head]>
+                >;
+              }
             : { [k in Head]: DeepPickFromPath<FlatModel[Head], Tail> }
           : never
         : Path extends keyof FlatModel
@@ -275,7 +279,7 @@ type ResolvedModel<
   Depth extends number = 7,
   RecursionLoop extends number[] = [-1, 0, 1, 2, 3, 4, 5, 6],
 > = {
-  done: NonRelationalFields<Model>;
+  done: NonRelationshipFields<Model>;
   recur: {
     [Field in keyof Model]: Model[Field] extends (
       ...args: any